+++ resources/js/Components/Status/StatusLists.js
@@ -1,9 +1,0 @@
-const StatusLists = [
-    "cancelled",
-    "research",
-    "whitepaper",
-    "pilot",
-    "development",
-    "launched",
-];
-export default StatusLists;